Create C# Learning Assignment: Summation of a series
Write a program in C# to find the sum of the series: 
 x - {x^3}/{3!} + {x^5}/{5!} - {x^7}/{7!} + {x^9}/{9!} - ...{x^n}/{n!} = sum
 
Here, the number of terms is specified by the user. Note that in this activity, there are spaces before and after the addition (+) or subtraction (-), and equal (=) operation. Finally, the sum will be followed by a newline.

Hint: Use Console.Write(String.Format("{0:0.00}", product));

Number of Terms 3, and x = 2
Number of Terms 7, and x = 4
2.00 - 1.33 + 0.27 = 0.93
4.00 - 10.67 + 8.53 - 3.25 + 0.72 - 0.11 + 0.03 = - 0.73