4 page website assignment - HTML and CSS only
The assignment was to create a responsive 4-page website using HTML and CSS. At the time of the assignment, we had not learned JS.
I chose to create a page about author and podcaster, Scott Sigler. He is one of the 3 original fiction podcasters who invented the podiobook concept.
Content IP on the page is copyright Scott Sigler, Empty Set Media, and Scott Pond. All IP content is used with permission, with my thanks to Scott Pond, A Sigler, and Scott Sigler for permission to create this project and post it here.